requirements.txt
Flask==0.12           
gunicorn==19.6.0
line-bot-sdk==1.5.0